feat(firebox#796): wasm64 source fixes for a buildable wasm64 sysroot
Source-only (regen'd api_*.h / __wasi*_real.c deliberately NOT committed — they
are 64-bit build artifacts that would break the wasm32 build; the firebox WASIX
overlay + a durable build.sh --wasm64 variant are a separate follow-up).

- arch/wasm64/pthread_arch.h: THE wasm64 TLS fix — read __wasilibc_pthread_self
  as a _Thread_local directly (match wasm32) instead of inline-asm global.get on
  a DATA symbol (which lld mis-resolved as a global-index reloc) + add __set_tp.
- mman/mman.c: place fixed-address mmaps (MAP_FIXED|MAP_FIXED_NOREPLACE) AT addr
  by growing the wasm linear memory — the linear-mapping requirement for a wasm64
  Blink guest. wasm32-safe (malloc path unchanged for non-fixed requests).
- sources/sbrk.c: optional __wasilibc_sbrk_max malloc-arena ceiling (default
  unlimited → no-op; an embedder confines its heap below a high linear region).
- sources/__wasilibc_poly_real.c: thread-spawn import returns int32_t (host tid),
  not size_t (wasm32-safe: int32_t==size_t there).
- pselect.c / ldso/dl*.c / thread/wasm64/wasi_thread_start.s: upstream wasm64
  width bugs (size_t vs __wasi_size_t; handle<->void* via uintptr_t; i32→i64 tls
  load). All wasm32-safe; upstream-PR candidates.
- tools/wasix-headers/src/c_header.rs: __wasi_size_t width-aware (uint64_t when
  mem_size_align==8).

+++ libc-bottom-half/sources/__wasilibc_poly_real.c
@@ -2,7 +2,13 @@
 #include <string.h>
 
 #ifdef _REENTRANT
-size_t __imported_wasi_thread_spawn(size_t arg0) __attribute__((
+// firebox#796: the `wasi.thread-spawn` host import returns a thread id (a fixed
+// i32 tid / -errno; see wasmer `thread_spawn<M> -> i32`), NOT a pointer-sized
+// value. The arg (start_arg pointer) is `size_t` = pointer-width (i64 on wasm64,
+// i32 on wasm32). The RETURN must be int32_t, else the wasm64 guest declares an
+// i64 result while the host gives i32 → `Instance::new` "incompatible import
+// type". int32_t == size_t on wasm32, so this is byte-identical there.
+int32_t __imported_wasi_thread_spawn(size_t arg0) __attribute__((
     __import_module__("wasi"),
     __import_name__("thread-spawn")
 ));
