+++ libc-top-half/musl/src/ldso/dlclose.c
@@ -4,7 +4,7 @@
 
 int dlclose(void *p)
 {
-	int err = __wasi_dl_invalid_handle((__wasi_dl_handle_t)p);
+	int err = __wasi_dl_invalid_handle((__wasi_dl_handle_t)(uintptr_t)p);
 	if (err != 0)
 	{
 		__dl_seterr("Invalid dl handle");
